+++ CHANGELOG.md
@@ -5,6 +5,11 @@ documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Changed
+- Improved container names, service containers names start with
+  `pipelines.<service>` instead of `pipelines-<service>`.
+
 ## [0.0.41] - 2020-06-21
 ### Added
 - Add `-c <name>=<value>` option to pass a configuration parameter to the